+++ test/tts_pipeline/test_phoneme.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import pytest
 
-from voicevox_engine.tts_pipeline.acoustic_feature_extractor import Phoneme
+from voicevox_engine.tts_pipeline.phoneme import Phoneme
 
 TRUE_NUM_PHONEME = 45
